requirements.txt
fastapi
uvicorn
maia2
python-chess